app/requirements.txt
streamlit
langchain-groq
langchain-community
langchain-core
python-dotenv
pandas
chromadb
PyPDF2
python-docx
uuid
